requirements.txt
# Core dependencies
python-binance==1.0.19
pandas
numpy
requests==2.31.0
ccxt>=4.0.0

# Dashboard
streamlit>=1.40.0
plotly==5.18.0

# Technical indicators
ta==0.11.0
pandas-ta-classic @ https://github.com/xgboosted/pandas-ta-classic/archive/refs/heads/main.zip

# Data storage
sqlalchemy==2.0.23

# Utilities
python-dotenv==1.0.0

# AI Integration
google-generativeai==0.3.2
anthropic>=0.18.0
